Fix options with scopes, minor refactoring
So, it turns out that `vim.bo` and `vim.wo` settings in `init.lua` ONLY
set the option for the very first open buffer.
Conversely, setting those same options with `vim.o` sets it for future
buffers, but not the initially open one.

This address it by abstracting setting options to a utility function
that sets `vim.o` as well as `vim.bo`/`vim.wo` options.

This also does some minor refactoring of other settings.