Remove git-blame in favor of gitsigns
Looks like gitsigns got more features since we first added it.
One of them duplicates what we were using git-blame for, so we can get
rid of that.

Also unifies the git keymaps into a single section per mode.
Added a keymap to preview what's changed in the git hunk.