Remove vim-go in favor of LSP, Treesitter features
vim-go was primarily kept for the auto-importing functionality, but with
Treesitter and LSP, we have all the other features.

By using `goimports` with `null-ls`, we can get rid of `vim-go`.
`vim-go` was not a bad plugin, but by using LSP and Treesitter, we can
unify language-specific settings around the same ecosystem of plugins,
rather than language-specific plugins, each with their own settings.